perf(davinci): skip legacy DOM transform on S2 build path
